+++ Source/KeyboardRespondable.swift
@@ -35,11 +35,15 @@ public extension KeyboardRespondable {
             return
         }
 
-        let convertedContentViewFrame = contentView.superview?.convert(contentView.frame, to: UIApplication.shared.keyWindow) ?? .zero
+        let convertedContentViewFrame =
+            contentView.superview?.convert(contentView.frame, to: UIApplication.shared.keyWindow) ?? .zero
         let contentViewMaxY = convertedContentViewFrame.maxY
         let keyboardMinY = keyboardFrame.minY
         let diff = max(0.0, contentViewMaxY - keyboardMinY)
 
-        contentView.contentInset = UIEdgeInsets(top: contentView.contentInset.top, left: contentView.contentInset.left, bottom: diff, right: contentView.contentInset.right)
+        contentView.contentInset = UIEdgeInsets(top: contentView.contentInset.top,
+                                                left: contentView.contentInset.left,
+                                                bottom: diff,
+                                                right: contentView.contentInset.right)
     }
 }